// src/data/dto/utilisateur/utilisateur-request.dto.ts
export interface UtilisateurRequestDTO {
  id: string;
  nom: string;
  prenom: string;
  imageUrl: string;
  telephone: string;
  email: string;
  ville: string;
  adresse: string;
}

/**
 * @typedef {object} UtilisateurRequestDTO
 * @property {string} id
 * @property {string} nom
 * @property {string} prenom
 * @property {string} telephone
 * @property {string} email
 * @property {string} ville
 * @property {string} adresse
 * @property {string} imageUrl
 */