// src/data/dto/utilisateur/utilisateur-response.dto.ts
export interface UtilisateurResponseDTO {
  id: string;
  nomPrenom: string;
  telephone: string;
  imageUrl: string;
  email: string;
  ville: string;
  adresse: string;
  dateInscription: string;
  dateDerniereConnexion: string;
}

/**
 * @typedef {object} UtilisateurResponseDTO
 * @property {string} id
 * @property {string} nomPrenom
 * @property {string} telephone
 * @property {string} imageUrl
 * @property {string} email
 * @property {string} ville
 * @property {string} adresse
 * @property {string} dateInscription
 * @property {string} dateDerniereConnexion
 */